Solution of workshop 3 exercise 1 (Correction)

+++ Practices/Workshop_3/src/Exercise_1.java
@@ -10,17 +10,17 @@ public class Exercise_1 {
     public static void main(String[] args) {
         Scanner entry = new Scanner(System.in);
         int n;
-        double sum = 0;
+        int sum = 0;
 
 
         System.out.println("Enter the vector size:");   //  We ask the user for the length of the array
         n = entry.nextInt();
 
-        double[] numbers = new double[n]; //  Create the array of N integers
+        int[] numbers = new int[n]; //  Create the array of N integers
 
         for (int i = 0; i < n ; i++) {  //   We iterate N times to go through the array positions and add values
             System.out.println("Enter the number " + (i+1) + ":");
-            numbers[i] = entry.nextDouble();
+            numbers[i] = entry.nextInt();
         }
 
         for (int i = 0; i < n ; i++) {  //   We use another FOR loop to sum up the array
